+++ index.js
@@ -288,7 +288,7 @@ Mapper.prototype._prepareReferences = function prepareReferences() {
     const {property, element} = r;
     switch (property) {
       case 'bpmn:sourceRef': {
-        const flow = this._upsertFlowRef(result, element.id, {
+        const flow = this._upsertRef(result.flowRefs, element.id, {
           id: element.id,
           $type: element.$type,
           sourceId: r.id,
@@ -299,7 +299,7 @@ Mapper.prototype._prepareReferences = function prepareReferences() {
         break;
       }
       case 'bpmn:targetRef': {
-        const flow = this._upsertFlowRef(result, element.id, {
+        const flow = this._upsertRef(result.flowRefs, element.id, {
           targetId: r.id,
         });
         const inbound = result.targetRefs[r.id] = result.targetRefs[r.id] || [];
@@ -307,7 +307,7 @@ Mapper.prototype._prepareReferences = function prepareReferences() {
         break;
       }
       case 'bpmn:default':
-        this._upsertFlowRef(result, r.id, {isDefault: true});
+        this._upsertRef(result.flowRefs, r.id, {isDefault: true});
         break;
       case 'bpmn:dataStoreRef':
         result.dataStoreRefs.push(r);
@@ -340,8 +340,8 @@ Mapper.prototype._prepareReferences = function prepareReferences() {
   return result;
 };
 
-Mapper.prototype._upsertFlowRef = function upsertFlowRef(result, id, value) {
-  const flow = result.flowRefs[id] = result.flowRefs[id] || {};
+Mapper.prototype._upsertRef = function upsertFlowRef(target, id, value) {
+  const flow = target[id] = target[id] || {};
   Object.assign(flow, value);
   return flow;
 };
@@ -553,7 +553,15 @@ Mapper.prototype._prepareActivity = function prepareActivity(element, parent, be
 };
 
 Mapper.prototype._prepareElementBehaviour = function prepareElementBehaviour(element, behaviour) {
-  const {id, $type: type, eventDefinitions, loopCharacteristics, ioSpecification, conditionExpression, properties} = element;
+  const {
+    id,
+    $type: type,
+    eventDefinitions,
+    loopCharacteristics,
+    ioSpecification,
+    conditionExpression,
+    properties,
+  } = element;
 
   const preparedElement = {
     ...behaviour,
@@ -593,6 +601,18 @@ Mapper.prototype._prepareElementBehaviour = function prepareElementBehaviour(ele
     preparedElement.ioSpecification = this._mapActivityBehaviour(ioSpecification, extendContext);
   }
 
+  if (element.dataInputAssociations) {
+    const associations = preparedElement.dataInputAssociations = [];
+    for (const association of element.dataInputAssociations) {
+      associations.push(this._mapActivityBehaviour(association, extendContext));
+    }
+  }
+  if (element.dataOutputAssociations) {
+    const associations = preparedElement.dataOutputAssociations = [];
+    for (const association of element.dataOutputAssociations) {
+      associations.push(this._mapActivityBehaviour(association, extendContext));
+    }
+  }
 
   if (conditionExpression && conditionExpression.language) {
     const {$type: exprType, language: scriptFormat, ...rest} = conditionExpression;
@@ -672,6 +692,13 @@ Mapper.prototype._mapActivityBehaviour = function mapActivityBehaviour(ed, exten
             value,
           });
         }
+      }
+      break;
+    }
+    case 'bpmn:DataOutputAssociation':
+    case 'bpmn:DataInputAssociation': {
+      if (Array.isArray(ed.sourceRef) && ed.sourceRef.length) {
+        behaviour.sourceRef = spreadRef(ed.sourceRef[0]);
       }
       break;
     }